% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invares.R
\name{check_invares}
\alias{check_invares}
\title{Performs checks on a fitted lavaan model for invariance effect size calculation}
\usage{
check_invares(fit, type = NULL)
}
\arguments{
\item{fit}{outputted multi-group cfa lavaan object}

\item{type}{character indicating which kind of check is to be performed ("lvmean" or "lvnum")}
}
\value{

}
\description{
Performs checks on a fitted lavaan model for invariance effect size calculation
}
