% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congeneric.R
\name{congeneric}
\alias{congeneric}
\title{A Function That Fits a Congeneric Reflective Latent Variable Measurement Model to a Set of Variables}
\usage{
congeneric(df, lvname)
}
\arguments{
\item{df}{input data frame consisting of variables in the measurement model}
}
\value{
a fitted lavaan object
}
\description{
This function fits a congeneric reflective latent variable model (described in McNeish and Wolf, 2020) using the indicators in the specified in the supplied data frame
and the lavaan function with fixed-factor scaling. FIML is used by default.
}
\examples{
cong.hs <- congeneric(df)
}
