% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorPcorProp.R
\name{CorPcorProp}
\alias{CorPcorProp}
\title{A Function That Calculates the Proportion of Correlations that are Stronger or Have a Different Sign than Partial Correlations}
\usage{
CorPcorProp(df, type = "obs", n = NULL, seed = NULL)
}
\arguments{
\item{df}{input data frame consisting of variables in the measurement model}

\item{type}{input character to (arbitrarily) name LV X in lavaan syntax. Defaults to "obs" for observed proportion but can simulate "ufm" (unidimensional factor model ) or "snm" (sparse network model)}

\item{n}{optional input number of simulated samples to use if specifying latent or network model. If comparing ufm and snm, the same
n should be specified.}

\item{seed}{optional input number to set random seed for simulations. If comparing ufm and snm, the same
seed should be specified.}
}
\value{
the observed proportion (for raw data) or a vector of simuluated proportions (based on specified model type)
}
\description{
This function calculates the proportion using either a raw data frame or a simulated common factor or network model. Based on procedure described by van Bork et al. (2019)
}
\examples{
obs <- CorPcorProp(df, type = "obs")
ufm <- CorPcorProp(df, type = "ufm", n = 100, seed = 123)
snm <- CorPcorProp(df, type = "snm", n = 100, seed = 123)
}
