% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfilters.R
\name{covfilters}
\alias{covfilters}
\title{A Function to Extract Covariate Levels for DIF Visualization}
\usage{
covfilters(df, covname, levs)
}
\arguments{
\item{df}{input data frame consisting of factor indicators and DIF covariate}

\item{covname}{input character vector name for DIF covariate in df}

\item{levs}{range of covariate levels for DIF examination (returned by covlevs)}
}
\value{
a list of filtered data frames consiting of the indicators of the factor (one per level of levs)
}
\description{
This function takes the outputted df from prods and generates the lavaan script necessary to fit a RFA model of uniform and non-uniform DIF
}
\examples{
filter.list <- covfilters(df, covname = "pers.lad", levs)
}
