% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covlevs.R
\name{covlevs}
\alias{covlevs}
\title{A Function to Extract Covariate Levels for DIF Visualization}
\usage{
covlevs(df, covname, covmin, covmax)
}
\arguments{
\item{df}{input data frame consisting of factor indicators and DIF covariate}

\item{covname}{input character vector name for DIF covariate in df}

\item{covmin}{input numeric for minimum value of covariate to begin estimating unidimensional factor models}

\item{covmax}{input numeric for maximum value of covariate to begin estimating unidimensional factor models}
}
\value{
a range of values within the covariate
}
\description{
This function takes the outputted df from prods and generates the lavaan script necessary to fit a RFA model of uniform and non-uniform DIF
}
\examples{
levs <- covlevs(df, covname = "pers.lad", covmin = 3, covmax = 8)
}
