% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difests.R
\name{difests}
\alias{difests}
\title{A Function that Extracts DIF-Relevant Parameters}
\usage{
difests(filter.dat, levs)
}
\arguments{
\item{filter.dat}{input list from covfilters}

\item{levs}{range of covariate levels for DIF examination (returned by covlevs)}
}
\value{
a data frame of parameter estimates from congeneric models fit on filtered dfs at each level of levs
}
\description{
This function takes the list from covfilters and runs a congeneric model on each and extracts and combines their parameter estimates
}
\examples{
outdf <- difests(filter.dat, levs)
}
