% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difviz.R
\name{difviz}
\alias{difviz}
\title{A Function that ggplots Uniform or Nonuniform DIF}
\usage{
difviz(ests, type = NULL, covtitle, lvtitle)
}
\arguments{
\item{ests}{output df of measurement model parameters from difests}

\item{type}{whether user wants to extract loading ("loads") or interpts "ints"}

\item{covtitle}{character input of name for covariate (both on x-axis title, and in file name)}

\item{lvtitle}{character input for name of lv (for file naming)}
}
\value{
a ggplot of uniform or nonuniform DIF
}
\description{
This function takes output from getdifests and creates a geom_tile viz to show uniform or non-uniform DIF. Rows for items that aren't flaged as showing sig. DIF are more translucent
}
\examples{
uni.plot <- difviz(ests, type = "ints", covtitle = "Social Class", lvtitle "Satisfaction")
}
