% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMultiMod.R
\name{fitMultiMod}
\alias{fitMultiMod}
\title{A function that scripts and fits a CFA model}
\usage{
fitMultiMod(df, vars, categ = FALSE)
}
\arguments{
\item{df}{the data frame that includes the to-be analyzed indicators}

\item{vars}{a named list of one or more character vectors of indicator names}

\item{categ}{an optional logical (defaults to FALSE) for whether indicators are categorical (in which case, ordered ULSMVS estimation is used, otherwise "MLR" is used)}
}
\value{

}
\description{
This function takes a named list of indicator variable names;
if there is only one character vector, a unidimensional CFA is fit,
but if there is more than one character vector, a multidimensional CFA
is fit, whereby the latent variables are named after the elements of the
list. Function by Steph Gauvin
}
\examples{
df <- psych::bfi

vars <- list(agree = c("A1","A2","A3","A4","A5"),
open = c("O1", "O2", "O3", "O4", "O5"),
neuro = c("N1", "N2", "N3", "N4", "N5"))

mod <- fitMultiMod(df, vars)
}
