% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invareffsize.R
\name{get_invareffsize}
\alias{get_invareffsize}
\title{User-facing function to calculate all noninvariance effect size measures in Gunn et al. (2020)}
\usage{
get_invareffsize(
  dat,
  fit,
  nodewidth = 0.01,
  lowerLV = -5,
  upperLV = 5,
  lvname = NULL
)
}
\arguments{
\item{dat}{data frame of indicators}

\item{fit}{outputted multi-group cfa lavaan object}

\item{nodewidth}{space between nodes during quadrature approximation (default = .01)}

\item{lowerLV}{lowest latent variable value evaluated (default = -5)}

\item{upperLV}{greatest latent variable value evaluated (default = 5)}

\item{lvname}{character to name outputted table and file.}

\item{table}{logical argument to request automated outputting of reproducible table (defaults to FALSE)}
}
\value{

}
\description{
User-facing function to calculate all noninvariance effect size measures in Gunn et al. (2020)
}
\examples{
dat <- HolzingerSwineford1939
dat$group[dat$school=="Pasteur"] = "Pasteur"
dat$group[dat$school=="Grant-White"] = "Grant-White"
HS.model <- '  visual =~ x1 + x2 + x3'
fit <- lavaan::cfa(HS.model, data = dat, group = "group")
}
