% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wsdi.R
\name{get_wsdi}
\alias{get_wsdi}
\title{Calculates weighted signed difference in expected indicator scores}
\usage{
get_wsdi(dat, fit, model, nodewidth = 0.01, lowerLV = -5, upperLV = 5)
}
\arguments{
\item{dat}{data frame of indicators}

\item{fit}{outputted multi-group cfa lavaan object}

\item{model}{inputted lavaan model syntax}

\item{nodewidth}{space between nodes during quadrature approximation (default = .01)}

\item{lowerLV}{lowest latent variable value evaluated (default = -5)}

\item{upperLV}{greatest latent variable value evaluated (default = 5)}
}
\value{

}
\description{
Calculates weighted signed difference in expected indicator scores
}
\examples{
#' dat <- HolzingerSwineford1939
dat$group[dat$school=="Pasteur"] = "Pasteur"
dat$group[dat$school=="Grant-White"] = "Grant-White"
HS.model <- '  visual =~ x1 + x2 + x3'
fit <- lavaan::cfa(HS.model, data = dat, group = "group")
}
