% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdifests.R
\name{getdifests}
\alias{getdifests}
\title{A Function that Creates a Filtered df for DIF plotting}
\usage{
getdifests(out, type = NULL, sig.uni = NULL, sig.nonuni = NULL)
}
\arguments{
\item{out}{output df of measurement model parameters from difests}

\item{type}{whether user wants to extract loading ("loads") or interpts "ints"}

\item{sig.uni}{an optional character vector of item #s that show significant uniform DIF}

\item{sig.nonuni}{an optional character vector of item #s that show significant non-uniform DIF}
}
\value{
a filtered data frame ready for ggplot
}
\description{
This function takes output from difests and filters it to a set of loadings or intercepts. User can specify a vector of items that show significant DIF for intercepts (uniform) or loadings (non-uniform) for later enhanced plotting
}
\examples{
ests <- getdifests(out, type = "loads", sig.nonuni = sig.nonuni)

}
