% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimicdif.R
\name{mimicdif}
\alias{mimicdif}
\title{A Function that Conducts a MIMIC model analysis of DIF}
\usage{
mimicdif(df, covname, lvname, match.op = FALSE, doubleMC.op = T, cat = FALSE)
}
\arguments{
\item{df}{input data frame consisting of variables in the measurement model}

\item{covname}{character input of name of covariate in data frame}

\item{lvname}{input character of latent varaible name for scripting}

\item{match.op}{supplies indprod (default to FALSE)}

\item{doubleMC.op}{supplies inprod (default to TRUE)}

\item{cat}{should a categorical estimator (estimator = "ULSMVS") be used? (default to FALSE)}
}
\value{
terminates with outputting reproducible table from mimicout
}
\description{
This function wraps the other MIMIC-related functions to fit and reproducibly
report on a MIMIC model analysis of DIF
}
\examples{
mimicdif(df, covname = "age", lvname = "openness")
}
