% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimicout.R
\name{mimicout}
\alias{mimicout}
\title{A function for reproducible MIMIC model output}
\usage{
mimicout(fit.mimic, mimic.param)
}
\arguments{
\item{fit.mimic}{fitted lavaan object from MIMIC script}

\item{mimic.param}{list of parameters generated from mimicparam}
}
\value{
outputs .doc to ./output subdirectory with reproducible output
}
\description{
This function takes the outputted lavaan MIMIC model and generates a reproducible table of output
}
\examples{
mimicout(fit.mimic, mimic.param)
}
