% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimicparam.R
\name{mimicparam}
\alias{mimicparam}
\title{A function that extracts list of MIMIC parameters to estimate}
\usage{
mimicparam(fit.mimic)
}
\arguments{
\item{fit.mimic}{fitted MIMIC model lavaan object}
}
\value{
a list of mimic parameters
}
\description{
This function takes the outputted lavaan MIMIC model and generates a reproducible table of output
}
\examples{
mimic.param <- mimicparam(fit.mimic)
}
