% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel}
\alias{parallel}
\title{A Function That Fits a Parallel Reflective Latent Variable Measurement Model to a Set of Variables}
\usage{
parallel(df)
}
\arguments{
\item{df}{input data frame consisting of variables in the measurement model}
}
\value{
a fitted lavaan object
}
\description{
This function fits a parallel (i.e., unit-weighting) reflective latent variable model (described in McNeish & Wolf, 2020) using the indicators  in the supplied data frame
and the lavaan function. Factor loadings are fixed to 1 and residual variances are fixed to equality
}
\examples{
par.hs <- parallel(df)
}
