% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.R
\name{pim}
\alias{pim}
\title{A Function That Fits a Pseudo-Indicator Model (PIM) to a Set of Variables}
\usage{
pim(df, model = "sum")
}
\arguments{
\item{df}{input data frame consisting of variables in the measurement model}

\item{model}{input character for whether measurement model to reproduce sum score (default) or average score is desired}
}
\value{
a fitted lavaan object
}
\description{
This function fits a PIM (Rose et al., 2019), to a set of indicators. Follows an adapted version of the
coding approach provided by McNeish and Wolf (2020, https://osf.io/7ea3k/)
}
\examples{
pim.hs <- pim(df, model = "average")
}
