% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{A Function that Calculates Coefficient H and Omega Total}
\usage{
reliability(model)
}
\arguments{
\item{model}{a fitted lavaan object from a measurement model for which the user wishes to estimate reliability}
}
\value{
a tibble of H and Omega Total
}
\description{
This function takes a fitted lavaan measurement model and calculates two indexes of latent reliability: Coefficient H and Omega Total. Based on McNeish (2018)
}
\examples{
cong.hs <- congeneric(df)
reliability(cong.hs)
}
