% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreit.R
\name{scoreit}
\alias{scoreit}
\title{A Function that Computes Factor Scores According to the Procedure Outlined By McNeish and Wolf}
\usage{
scoreit(itemdf, datadf, scorerole = "NULL")
}
\arguments{
\item{itemdf}{input data frame of the items used when executing psyscores::congeneric()}

\item{datadf}{target data frame to which the computed vector of factor scores should be appended}

\item{scorerole}{input character for which analytic role the factor score will play. When "predictor"
is specified, regression/MAP scores with Croon (2002) correction are computed. When "outcome" is specified,
Bartlett scores are computed. No default to prevent accidental specification.}
}
\value{
data frame consisting of data frame specified in datadf with the additional column of factor scores
}
\description{
This function takes the df of items and executes psyscores::congeneric().
It then takes the lavaan model and applies the procedure outlined by McNeish and Wolf (2020), following from Skrondal & Laake (2001)
}
\examples{
dvn = dyadVarNames(dat, xvar="X", sep = ".",distinguish1 = "1", distinguish2 = "2")
}
