% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_stdmod.R
\name{scrape_stdmod}
\alias{scrape_stdmod}
\title{Scrape standardized model statement from fitted lavaan object}
\usage{
scrape_stdmod(fit)
}
\arguments{
\item{fit}{outputted \code{\link{lavaan}} object}
}
\value{

}
\description{
Scrape standardized model statement from fitted lavaan object
}
